requirements.txt
flask
flask_restful
langchain
langchain-community
langchain_google_genai
python-dotenv

